// typescript/singleAsset.interface.ts
export interface SingleAsset {
    data: Data
    timestamp: number
  }
  
  export interface Data {
    id: string
    rank: string
    explorer: string
    symbol: string
    name: string
    supply: string
    maxSupply: string
    marketCapUsd: string
    volumeUsd24Hr: string
    priceUsd: string
    changePercent24Hr: string
    vwap24Hr: string
  }
  